Created a new class for the address parser that holds all the pre- and post-processing steps.
